Use a query param to turn on fixture
