+++ package.json
@@ -16,9 +16,11 @@
     "@types/underscore": "^1.9.4",
     "moment": "^2.24.0",
     "papaparse": "^5.1.1",
+    "query-string": "^6.12.1",
     "react": "^16.12.0",
     "react-csv": "^2.0.1",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.1",
     "react-table": "^7.0.0",
     "react-tabs": "^3.1.0",
